refactor(calendar): consolidate types, fix to-date inclusivity, tighten code
- Move CalendarEvent + CALENDAR_EVENT_TYPES to @atrium/shared so API and web
  share one source of truth instead of two parallel definitions.
- Fix end-of-day handling on the `to` query param: a date-only string parsed
  to 00:00Z, silently excluding any task/invoice timestamped later that day.
- Drop the duplicate CalendarQuery interface in the service; use
  CalendarQueryDto directly.
- Use TASK_STATUSES constants in event-chip instead of stringly-typed
  comparisons.
- Fix gridDays DST hazard by constructing each Date from indices rather
  than mutating with setDate.
- Memoize byDate Map and gridDays in month-grid; memoize grouping in
  agenda-list. Add shared groupByDate helper to remove duplication.
- Add AbortController to the calendar page load so rapid month navigation
  cannot land stale data; drop the duplicate error toast (inline retry
  block already surfaces failures).
- Hoist getOrCreateProject and createTask into e2e/tests/helpers.ts;
  remove dead dismissTelemetry from the calendar e2e (no other suite
  uses it and the banner is not present in seeded test accounts).
- Spec data: tasks now use status: "open" to match TASK_STATUSES.
